+++ include/nanobind/nb_cast.h
@@ -126,10 +126,23 @@ template <typename T>
 struct type_caster<T, enable_if_t<std::is_arithmetic_v<T> && !is_std_char_v<T>>> {
     NB_INLINE bool from_python(handle src, uint8_t flags, cleanup_list *) noexcept {
         if constexpr (std::is_floating_point_v<T>) {
-            if constexpr (sizeof(T) == 8)
+            if constexpr (std::is_same_v<T, double>) {
                 return detail::load_f64(src.ptr(), flags, &value);
-            else
+            } else if constexpr (std::is_same_v<T, float>) {
                 return detail::load_f32(src.ptr(), flags, &value);
+            } else {
+                double d;
+                if (!detail::load_f64(src.ptr(), flags, &d))
+                    return false;
+                T result = (T) d;
+                if ((flags & (uint8_t) cast_flags::convert)
+                        || (double) result == d
+                        || (result != result && d != d)) {
+                    value = result;
+                    return true;
+                }
+                return false;
+            }
         } else {
             if constexpr (std::is_signed_v<T>) {
                 if constexpr (sizeof(T) == 8)